+++ AccessibilityBundles/PhotoLibraryFramework.axbundle/PLImageAccessibilityElement.h
@@ -3,19 +3,19 @@
  */
 
 @interface PLImageAccessibilityElement : UIAccessibilityElement {
-    NSInteger _index;
+    int _index;
 }
 
-@property NSInteger index;
+@property int index;
 
-- (NSInteger)_albumPhotoIndex;
+- (int)_albumPhotoIndex;
 - (struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })accessibilityFrame;
 - (id)accessibilityHint;
 - (id)accessibilityLabel;
 - (unsigned long long)accessibilityTraits;
 - (id)dataSource;
 - (id)delegate;
-- (NSInteger)index;
-- (void)setIndex:(NSInteger)arg1;
+- (int)index;
+- (void)setIndex:(int)arg1;
 
 @end
