+++ AccessibilityBundles/TelephonyUIFramework.axbundle/AXTelephonyUIGlue.h
@@ -7,8 +7,8 @@
 
 + (void)accessibilityInitializeBundle;
 
-- (NSInteger)_keyForPoint:(struct CGPoint { float x1; float x2; })arg1;
-- (struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })_rectForKey:(NSInteger)arg1;
+- (int)_keyForPoint:(struct CGPoint { float x1; float x2; })arg1;
+- (struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })_rectForKey:(int)arg1;
 - (void)unlock;
 
 @end
