+++ AccessibilityBundles/UIKit.axbundle/UIAlertViewAccessibility.h
@@ -9,10 +9,10 @@
 
 - (void)_accessibilitySendAlert;
 - (void)_performPopoutAnimationAnimated:(BOOL)arg1;
-- (void)_performPopup:(BOOL)arg1 animationType:(NSInteger)arg2;
+- (void)_performPopup:(BOOL)arg1 animationType:(int)arg2;
 - (void)_setTextFieldsHidden:(BOOL)arg1;
 - (void)_slideSheetOut:(BOOL)arg1;
-- (void)setCancelButtonIndex:(NSInteger)arg1;
+- (void)setCancelButtonIndex:(int)arg1;
 - (void)show;
 
 @end
