+++ AccessibilityBundles/UIKit.axbundle/UISearchBarTextFieldAccessibility.h
@@ -9,6 +9,6 @@
 
 - (id)accessibilityLabel;
 - (unsigned long long)accessibilityTraits;
-- (BOOL)textField:(id)arg1 shouldChangeCharactersInRange:(struct _NSRange { NSUInteger x1; NSUInteger x2; })arg2 replacementString:(id)arg3;
+- (BOOL)textField:(id)arg1 shouldChangeCharactersInRange:(struct _NSRange { unsigned int x1; unsigned int x2; })arg2 replacementString:(id)arg3;
 
 @end
