+++ Frameworks/AVFoundation.framework/AVAVAudioSettingsAudioOutputSettings.h
@@ -10,8 +10,8 @@
 + (id)eligibleOutputSettingsDictionaryKeys;
 
 - (id)audioOptions;
-- (struct AudioChannelLayout { NSUInteger x1; NSUInteger x2; NSUInteger x3; struct AudioChannelDescription { NSUInteger x_4_1_1; NSUInteger x_4_1_2; float x_4_1_3[3]; } x4[1]; }*)copyAudioChannelLayoutAndSize:(NSUInteger*)arg1;
-- (void)getAudioStreamBasicDescription:(struct AudioStreamBasicDescription { double x1; NSUInteger x2; NSUInteger x3; NSUInteger x4; NSUInteger x5; NSUInteger x6; NSUInteger x7; NSUInteger x8; NSUInteger x9; }*)arg1 forAudioFileTypeID:(unsigned long)arg2 sourceFormatDescription:(struct opaqueCMFormatDescription { }*)arg3;
+- (struct AudioChannelLayout { unsigned int x1; unsigned int x2; unsigned int x3; struct AudioChannelDescription { unsigned int x_4_1_1; unsigned int x_4_1_2; float x_4_1_3[3]; } x4[1]; }*)copyAudioChannelLayoutAndSize:(unsigned int*)arg1;
+- (void)getAudioStreamBasicDescription:(struct AudioStreamBasicDescription { double x1; unsigned int x2; unsigned int x3; unsigned int x4; unsigned int x5; unsigned int x6; unsigned int x7; unsigned int x8; unsigned int x9; }*)arg1 forAudioFileTypeID:(unsigned long)arg2 sourceFormatDescription:(struct opaqueCMFormatDescription { }*)arg3;
 - (id)initWithAVAudioSettingsDictionary:(id)arg1 exceptionReason:(id*)arg2;
 - (BOOL)isDictionaryFullyFormed;
 - (BOOL)willYieldCompressedSamples;
