+++ Frameworks/AVFoundation.framework/AVAssetTrackGroup.h
@@ -9,11 +9,11 @@
     NSArray *_trackIDs;
 }
 
-@property(readonly) NSArray *trackIDs;
+@property(readonly) NSArray * trackIDs;
 
 - (id)_assetComparisonToken;
 - (void)dealloc;
-- (NSUInteger)hash;
+- (unsigned int)hash;
 - (id)init;
 - (id)initWithAsset:(id)arg1 trackIDs:(id)arg2;
 - (BOOL)isEqual:(id)arg1;
